style(emails): code style tweaks
* moved email-related features to services/email
* enforced fonts in whole mail
* added some tsdoc

What have not been done yet
* docs/api
* split defined mails into separate files
* decide whether to wait for mail to be successfully sent
