+++ assets/email/template.html
@@ -56,7 +56,7 @@
                   <td>
                     <span
                       style="
-                        font-family: Montserrat;
+                        font-family: Montserrat, sans-serif;
                         color: #fff;
                         font-size: 22px;
                         padding: 20px;
@@ -79,7 +79,12 @@
                               <tbody>
                                 <tr>
                                   <td
-                                    style="font-family: Nunito; line-height: 21px; font-size: 14px; text-align: center"
+                                    style="
+                                      font-family: Nunito, Verdana, Geneva, Tahoma, sans-serif;
+                                      line-height: 21px;
+                                      font-size: 14px;
+                                      text-align: center;
+                                    "
                                   >
                                     {{title.short}}
                                   </td>
@@ -91,7 +96,7 @@
                                       line-height: 36px;
                                       font-size: 30px;
                                       font-style: normal;
-                                      font-family: Montserrat;
+                                      font-family: Montserrat, sans-serif;
                                       color: #202020;
                                     "
                                   >
@@ -113,7 +118,7 @@
                                 <tr>
                                   <td
                                     style="
-                                      font-family: Montserrat;
+                                      font-family: Montserrat, sans-serif;
                                       line-height: 29px;
                                       font-size: 24px;
                                       color: #006492;
@@ -139,7 +144,7 @@
                                 <tr>
                                   <td
                                     style="
-                                      font-family: Montserrat;
+                                      font-family: Montserrat, sans-serif;
                                       line-height: 29px;
                                       font-size: 24px;
                                       color: #006492;
@@ -155,7 +160,7 @@
                                     -webkit-text-size-adjust: none;
                                     -ms-text-size-adjust: none;
                                     position: relative;
-                                    font-family: Nunito;
+                                    font-family: Nunito, Verdana, Geneva, Tahoma, sans-serif;
                                   "
                                 >
                                   <td>Le contenu du mail n'a pas été chargé !</td>
@@ -166,7 +171,7 @@
                                     -webkit-text-size-adjust: none;
                                     -ms-text-size-adjust: none;
                                     position: relative;
-                                    font-family: Nunito;
+                                    font-family: Nunito, Verdana, Geneva, Tahoma, sans-serif;
                                   "
                                 >
                                   <td>
@@ -206,7 +211,7 @@
                               text-align: center;
                               -webkit-text-size-adjust: none;
                               -ms-text-size-adjust: none;
-                              font-family: Nunito;
+                              font-family: Nunito, Verdana, Geneva, Tahoma, sans-serif;
                             "
                           >
                             Suivez-nous sur les réseaux sociaux pour suivre les infos en temps réel pendant
@@ -353,7 +358,7 @@
                               text-align: center;
                               -webkit-text-size-adjust: none;
                               -ms-text-size-adjust: none;
-                              font-family: Nunito;
+                              font-family: Nunito, Verdana, Geneva, Tahoma, sans-serif;
                             "
                           >
                             Ce mail a été envoyé à {{receiver}}
@@ -371,7 +376,7 @@
                               text-align: center;
                               -webkit-text-size-adjust: none;
                               -ms-text-size-adjust: none;
-                              font-family: Nunito;
+                              font-family: Nunito, Verdana, Geneva, Tahoma, sans-serif;
                             "
                           >
                             {{reason}}
@@ -387,7 +392,7 @@
                               text-align: center;
                               -webkit-text-size-adjust: none;
                               -ms-text-size-adjust: none;
-                              font-family: Nunito;
+                              font-family: Nunito, Verdana, Geneva, Tahoma, sans-serif;
                             "
                           >
                             © {{& year}}
@@ -409,7 +414,7 @@
                               text-align: center;
                               -webkit-text-size-adjust: none;
                               -ms-text-size-adjust: none;
-                              font-family: Nunito;
+                              font-family: Nunito, Verdana, Geneva, Tahoma, sans-serif;
                             "
                           >
                             12 rue Marie Curie, 10000 Troyes
